+++ sql/02-create-tuwel.sql
@@ -33,8 +33,8 @@ CREATE TABLE tuwel.course
 
 CREATE TABLE tuwel.event
 (
-    event_id    INT  NOT NULL,
-    course_id   INT  NOT NULL,
+    event_id    BIGINT NOT NULL,
+    course_id   INT    NOT NULL,
 
     start_ts    TIMESTAMP WITH TIME ZONE,
     end_ts      TIMESTAMP WITH TIME ZONE,
@@ -42,7 +42,7 @@ CREATE TABLE tuwel.event
     access_ts   timestamp WITH TIME ZONE,
     mod_ts      TIMESTAMP WITH TIME ZONE,
 
-    name        TEXT NOT NULL,
+    name        TEXT   NOT NULL,
     description TEXT,
 
     CONSTRAINT pk_event PRIMARY KEY (event_id),
@@ -55,8 +55,8 @@ CREATE INDEX start_ts ON tuwel.event (start_ts);
 
 CREATE TABLE tuwel.event_user
 (
-    event_id INT NOT NULL,
-    user_id  INT NOT NULL,
+    event_id BIGINT NOT NULL,
+    user_id  INT    NOT NULL,
 
     CONSTRAINT pk_event_user PRIMARY KEY (event_id, user_id),
     CONSTRAINT fk_event_user_event FOREIGN KEY (event_id) REFERENCES tuwel.event (event_id)
